+++ Data/Void/Unsafe.hs
@@ -5,8 +5,7 @@
 #endif
 -----------------------------------------------------------------------------
 -- |
--- Module      :  Data.Void.Unsafe
--- Copyright   :  (C) 2008-2012 Edward Kmett
+-- Copyright   :  (C) 2008-2013 Edward Kmett
 -- License     :  BSD-style (see the file LICENSE)
 --
 -- Maintainer  :  Edward Kmett <ekmett@gmail.com>